refactor: move 'settings' vuex store to 'display' store
