+++ apps/web/src/schema/cost-analysis/cost-report-config/api-verbs/update.ts
@@ -1,4 +1,4 @@
-import type { Currency } from '@/store/modules/settings/type';
+import type { Currency } from '@/store/modules/display/type';
 import type { LanguageCode } from '@/store/modules/user/type';
