+++ apps/web/src/services/cost-explorer/components/CostReportSettingsModal.vue
@@ -14,8 +14,8 @@ import type { CostReportConfigUpdateParameters } from '@/schema/cost-analysis/co
 import type { CostReportConfigModel } from '@/schema/cost-analysis/cost-report-config/model';
 import { i18n } from '@/translations';
 
-import { CURRENCY, CURRENCY_SYMBOL } from '@/store/modules/settings/config';
-import type { Currency } from '@/store/modules/settings/type';
+import { CURRENCY, CURRENCY_SYMBOL } from '@/store/modules/display/config';
+import type { Currency } from '@/store/modules/display/type';
 import { languages } from '@/store/modules/user/config';
 import type { LanguageCode } from '@/store/modules/user/type';
