+++ apps/web/src/services/dashboards/widgets/_components/WidgetDataTable.vue
@@ -12,8 +12,8 @@ import {
 } from '@cloudforet/mirinae';
 import { numberFormatter, getValueByPath } from '@cloudforet/utils';
 
-import { CURRENCY } from '@/store/modules/settings/config';
-import type { Currency } from '@/store/modules/settings/type';
+import { CURRENCY } from '@/store/modules/display/config';
+import type { Currency } from '@/store/modules/display/type';
 
 import { currencyMoneyFormatter } from '@/lib/helper/currency-helper';
 import { usageUnitFormatter } from '@/lib/helper/usage-formatter';
