+++ apps/web/src/services/dashboards/widgets/_components/WidgetFrame.vue
@@ -17,8 +17,8 @@ import type { DateRange } from '@/schema/dashboard/_types/dashboard-type';
 import type { Granularity, WidgetSize } from '@/schema/dashboard/_types/widget-type';
 import { i18n } from '@/translations';
 
-import { CURRENCY_SYMBOL } from '@/store/modules/settings/config';
-import type { Currency } from '@/store/modules/settings/type';
+import { CURRENCY_SYMBOL } from '@/store/modules/display/config';
+import type { Currency } from '@/store/modules/display/type';
 
 import getRandomId from '@/lib/random-id-generator';
