+++ apps/web/src/store/index.ts
@@ -3,7 +3,6 @@ import Vuex from 'vuex';
 
 import display from './modules/display';
 import domain from './modules/domain';
-import settings from './modules/settings';
 // eslint-disable-next-line import/no-cycle
 import user from './modules/user';
 // eslint-disable-next-line import/no-cycle
@@ -14,7 +13,6 @@ Vue.use(Vuex);
 const store = new Vuex.Store<any>({
     modules: {
         user,
-        settings,
         domain,
         display,
     },
