+++ apps/web/src/store/modules/display/config.ts
@@ -4,3 +4,15 @@ export const SIDEBAR_TYPE = Object.freeze({
     widget: 'widget',
 });
 export type SidebarType = typeof SIDEBAR_TYPE[keyof typeof SIDEBAR_TYPE];
+
+export const CURRENCY = Object.freeze({
+    USD: 'USD',
+    KRW: 'KRW',
+    JPY: 'JPY',
+});
+
+export const CURRENCY_SYMBOL = Object.freeze({
+    USD: '$',
+    KRW: '₩',
+    JPY: '¥',
+});
