+++ apps/web/src/store/modules/display/type.ts
@@ -1,11 +1,14 @@
 import type { TranslateResult } from 'vue-i18n';
 import type { Location } from 'vue-router';
 
-import type { SidebarType } from '@/store/modules/display/config';
+import type { SidebarType, CURRENCY, CURRENCY_SYMBOL } from '@/store/modules/display/config';
 
 import type { Menu } from '@/lib/menu/config';
 
+
 export type HighlightTagType = 'new' | 'beta' | 'update';
+export type Currency = typeof CURRENCY[keyof typeof CURRENCY];
+export type CurrencySymbol = typeof CURRENCY_SYMBOL[keyof typeof CURRENCY_SYMBOL];
 
 export interface DisplayMenu extends Menu {
     show?: boolean;
@@ -25,6 +28,7 @@ export interface DisplayState {
     uncheckedNoticeCount: number,
     isSignInFailed: boolean;
     visibleMobileGuideModal: boolean;
+    gnbNotificationLastReadTime: string;
 }
 
 export interface SidebarProps {
