--- apps/web/src/store/modules/settings/config.ts
@@ -1,11 +1,0 @@
-export const CURRENCY = Object.freeze({
-    USD: 'USD',
-    KRW: 'KRW',
-    JPY: 'JPY',
-});
-
-export const CURRENCY_SYMBOL = Object.freeze({
-    USD: '$',
-    KRW: '₩',
-    JPY: '¥',
-});
