--- apps/web/src/store/modules/settings/index.ts
@@ -1,14 +1,0 @@
-import * as actions from './actions';
-import * as mutations from './mutations';
-import type { SettingsState } from './type';
-
-const state: SettingsState = {
-    gnbNotificationLastReadTime: '',
-};
-
-export default {
-    namespaced: true,
-    state,
-    actions,
-    mutations,
-};
