--- apps/web/src/store/modules/settings/type.ts
@@ -1,13 +1,0 @@
-import type { CURRENCY, CURRENCY_SYMBOL } from '@/store/modules/settings/config';
-
-export type Currency = typeof CURRENCY[keyof typeof CURRENCY];
-export type CurrencySymbol = typeof CURRENCY_SYMBOL[keyof typeof CURRENCY_SYMBOL];
-
-export type CurrencyRates = {
-    [K in Currency]: number;
-};
-
-export interface SettingsState {
-    gnbNotificationLastReadTime: string;
-}
-
