+++ apps/web/src/store/plugins.ts
@@ -1,6 +1,6 @@
 import createLogger from 'vuex/dist/logger';
 
-import settingsPlugins from '@/store/modules/settings/plugins';
+import displayPlugins from '@/store/modules/display/plugins';
 // eslint-disable-next-line import/no-cycle
 import userPlugins from '@/store/modules/user/plugins';
 
@@ -8,7 +8,7 @@ const DEBUG = import.meta.env.DEV;
 
 const PLUGINS = [
     ...userPlugins,
-    ...settingsPlugins,
+    ...displayPlugins,
 ];
 
 const logger = createLogger({
